Move DatasetGenerator and ModelParser to dsl package
