+++ core/src/test/scala/latis/ops/CurrySuite.scala
@@ -7,8 +7,8 @@ import org.scalatest.matchers.should.Matchers._
 import latis.data._
 import latis.dataset.Dataset
 import latis.dataset.MemoizedDataset
+import latis.dsl._
 import latis.model._
-import latis.util.DatasetGenerator
 
 class CurrySuite extends AnyFunSuite {
