+++ core/src/test/scala/latis/ops/DropRightSpec.scala
@@ -9,7 +9,6 @@ import latis.data.Sample
 import latis.data.SampledFunction
 import latis.dataset.Dataset
 import latis.dsl._
-import latis.util.DatasetGenerator
 
 class DropRightSpec extends AnyFlatSpec {
