+++ core/src/test/scala/latis/ops/EvaluationSpec.scala
@@ -5,7 +5,6 @@ import org.scalatest.matchers.should.Matchers._
 
 import latis.data._
 import latis.dsl._
-import latis.util.DatasetGenerator
 
 class EvaluationSpec extends AnyFlatSpec {
