+++ core/src/test/scala/latis/ops/PivotSuite.scala
@@ -6,8 +6,8 @@ import org.scalatest.matchers.should.Matchers._
 
 import latis.data._
 import latis.dataset.Dataset
+import latis.dsl._
 import latis.model._
-import latis.util.DatasetGenerator
 
 class PivotSuite extends AnyFunSuite {
